refactor: rename "simple" to "single"
