+++ .prettierignore
@@ -1,7 +1,7 @@
 lib/
 pnpm-lock.yaml
 
-src/nodes/typeGuards/simple.ts
+src/nodes/typeGuards/single.ts
 src/nodes/typeGuards/union.ts
 
 # See https://github.com/all-contributors/cli/issues/347
