+++ src/nodes/typeGuards/single.ts
